refactor(host): split @lace/host so its barrel is vscode-free
Move vscode-coupled files out of @lace/host's published surface. They
weren't library code — they're the extension shell's activation +
webview-panel surface, used only by this repo.

Moves:
- packages/host/src/extension.ts          → src/extension.ts
- packages/host/src/canvas-panel.ts       → src/vscode/canvas-panel.ts
- packages/host/src/module-detail-panel.ts → src/vscode/module-detail-panel.ts
- packages/host/src/registry-sidebar.ts   → src/vscode/registry-sidebar.ts
- packages/host/src/webview-html.ts       → src/vscode/webview-html.ts
- handleRpcError (the vscode.window toast-shower) extracted from
  packages/host/src/rpc-errors.ts → src/vscode/rpc-error-ui.ts

What stays in @lace/host: LaceTransport, ServerManager, SubscribeHandler,
engine-process, RpcError + classifyRpcError + requireClient (the pure
classifier — no vscode), proto-gen types. These are the bits future
portal/JetBrains adapters will reuse.

ChatViewProvider is now constructed with a `BuildWebviewHtml`
callback supplied by the extension shell, not by importing
buildWebviewHtml from @lace/host. Inversion of control — the package
stays vscode-host-free internally, the JetBrains adapter will supply
its own builder.

Delete vitest.setup.ts and drop setupFiles from vitest.config.ts.
After the split, no vscode-free package transitively loads vscode at
import time, so the global vi.mock('vscode', …) shim is no longer
needed. If a future import inadvertently re-introduces the coupling,
the test crashes — that's the signal we want, not a defensive shim
that masks it. Greenfield posture.

All 126 unit tests pass with no ambient vscode mock (setup: 0ms).
3 rspack bundles compile clean. Biome + tsc green.
